requirements-dev.txt
Pillow
matplotlib
numpy
requests
tqdm
